fix(jazz-tools): wire ConnectSyncRuntimeOptions.onAuthFailure to runtime.onAuthFailure
Extracts mapAuthReason from jazz-worker.ts to auth-state.ts (shared), adds
onAuthFailure? to the Runtime interface, and wires the callback in the
JazzClient constructor so callers of connectWithRuntime/connect/connectSync
receive typed auth-failure notifications from the Rust transport.
